Created a chessLogic file
Created a chessLogic file.
This is the file, where the functions required for computation of moves made by a computer will be stored.
By now only working function there is a static evaluation function, that simply evaluates the position by counting the material advantage of a white player.
Also made some flicks in a chessEngine file.